+++ lib/src/assets_audio_player.dart
@@ -51,6 +51,7 @@ const String METHOD_NOTIFICATION_PREV = 'player.prev';
 const String METHOD_NOTIFICATION_STOP = 'player.stop';
 const String METHOD_NOTIFICATION_PLAY_OR_PAUSE = 'player.playOrPause';
 const String METHOD_PLAY_SPEED = 'player.playSpeed';
+const String METHOD_PITCH = 'player.pitch';
 const String METHOD_ERROR = 'player.error';
 const String METHOD_AUDIO_SESSION_ID = 'player.audioSessionId';
 const String METHOD_PLAY_AFTER_INTERRUPTION = 'player.afterInterruption';
@@ -91,8 +92,8 @@ class PlayerEditor {
   void onAudioReplacedAt(int index, bool keepPlayingPositionIfCurrent) {
     assetsAudioPlayer._updatePlaylistIndexes();
     if (assetsAudioPlayer._playlist!.playlistIndex == index) {
-      final currentPosition = assetsAudioPlayer.currentPosition.value;
-      final isPlaying = assetsAudioPlayer.isPlaying.value ?? false;
+      final currentPosition = assetsAudioPlayer.currentPosition.valueOrNull;
+      final isPlaying = assetsAudioPlayer.isPlaying.valueOrNull ?? false;
       //print('onAudioReplacedAt/ currentPosition : $currentPosition');
       if (keepPlayingPositionIfCurrent && currentPosition != null) {
         assetsAudioPlayer._openPlaylistCurrent(
@@ -133,6 +134,9 @@ class AssetsAudioPlayer {
   static final double maxPlaySpeed = 16.0;
   static final double defaultVolume = maxVolume;
   static final double defaultPlaySpeed = 1.0;
+  static final double minPitch = 0.0;
+  static final double maxPitch = 16.0;
+  static final double defaultPitch = 1.0;
   static final AudioFocusStrategy defaultFocusStrategy =
       AudioFocusStrategy.request(resumeAfterInterruption: true);
   static final NotificationSettings defaultNotificationSettings =
@@ -425,6 +429,10 @@ class AssetsAudioPlayer {
 
   ValueStream<double> get playSpeed => _playSpeed.stream;
 
+  final BehaviorSubject<double> _pitch = BehaviorSubject.seeded(1.0);
+
+  ValueStream<double> get pitch => _pitch.stream;
+
   final BehaviorSubject<double> _forwardRewindSpeed = BehaviorSubject.seeded(0);
 
   ValueStream<double> get forwardRewindSpeed => _forwardRewindSpeed.stream;
@@ -434,7 +442,7 @@ class AssetsAudioPlayer {
   /// returns the looping state : true -> looping, false -> not looping
   LoopMode? get currentLoopMode => _loopMode.value;
 
-  bool get shuffle => _shuffle.value ?? false;
+  bool get shuffle => _shuffle.valueOrNull ?? false;
 
   bool _stopped = false;
 
@@ -578,7 +586,7 @@ class AssetsAudioPlayer {
           break;
         case METHOD_CURRENT:
           if (call.arguments == null) {
-            final current = _current.value;
+            final current = _current.valueOrNull;
             if (current != null) {
               final finishedPlay = Playing(
                 audio: current.audio,
@@ -638,6 +646,9 @@ class AssetsAudioPlayer {
         case METHOD_PLAY_SPEED:
           _playSpeed.add(call.arguments);
           break;
+        case METHOD_PITCH:
+          _pitch.add(call.arguments);
+          break;
         case METHOD_FORWARD_REWIND_SPEED:
           final double newValue = call.arguments;
           if (_forwardRewindSpeed.value != newValue) {
@@ -672,7 +683,7 @@ class AssetsAudioPlayer {
             pause();
             break;
           case PlayInBackground.disabledRestoreOnForeground:
-            _wasPlayingBeforeEnterBackground = isPlaying.value ?? false;
+            _wasPlayingBeforeEnterBackground = isPlaying.valueOrNull ?? false;
             pause();
             break;
         }
@@ -746,8 +757,8 @@ class AssetsAudioPlayer {
   Future<bool> previous({bool keepLoopMode = true}) async {
     if (_playlist != null) {
       // more than 5 sec played, go back to the start of audio
-      if (_currentPosition.value != null &&
-          _currentPosition.value!.inSeconds >= 5) {
+      if (_currentPosition.valueOrNull != null &&
+          _currentPosition.valueOrNull!.inSeconds >= 5) {
         await seek(Duration.zero, force: true);
       } else if (_playlist!.hasPrev()) {
         if (!keepLoopMode) {
@@ -768,7 +779,7 @@ class AssetsAudioPlayer {
   }
 
   void _onPositionReceived(dynamic argument) {
-    final oldValue = _currentPosition.value;
+    final oldValue = _currentPosition.valueOrNull;
     int? newValue;
     if (argument is int) {
       final value = argument;
@@ -822,6 +833,7 @@ class AssetsAudioPlayer {
         respectSilentMode: _playlist!.respectSilentMode,
         showNotification: _playlist!.showNotification,
         playSpeed: _playlist!.playSpeed,
+        pitch: _playlist!.pitch,
         notificationSettings: _playlist!.notificationSettings,
         autoStart: autoStart,
         loopMode: _playlist!.loopMode,
@@ -865,7 +877,7 @@ class AssetsAudioPlayer {
         }
       }
       if (_playlist!.hasNext()) {
-        final curr = _current.value;
+        final curr = _current.valueOrNull;
         if (curr != null) {
           _playlistAudioFinished.add(Playing(
             audio: curr.audio,
@@ -880,7 +892,7 @@ class AssetsAudioPlayer {
         return true;
       } else if (loopMode.value == LoopMode.playlist) {
         //last element
-        final curr = _current.value;
+        final curr = _current.valueOrNull;
         if (curr != null) {
           _playlistAudioFinished.add(Playing(
             audio: curr.audio,
@@ -899,7 +911,7 @@ class AssetsAudioPlayer {
         return true;
       } else if (requestByUser) {
         //last element
-        final curr = _current.value;
+        final curr = _current.valueOrNull;
         if (curr != null) {
           _playlistAudioFinished.add(Playing(
             audio: curr.audio,
@@ -1013,17 +1025,18 @@ class AssetsAudioPlayer {
   // private method, used in open(playlist) and open(path)
   Future<void> _open(
     Audio? audioInput, {
-    bool? autoStart,
-    double? forcedVolume,
-    bool? respectSilentMode,
-    bool? showNotification,
-    Duration? seek,
-    double? playSpeed,
-    LoopMode? loopMode,
-    HeadPhoneStrategy? headPhoneStrategy,
-    AudioFocusStrategy? audioFocusStrategy,
-    NotificationSettings? notificationSettings,
-    bool? needRecord = false,
+    required bool? autoStart,
+    required double? forcedVolume,
+    required bool? respectSilentMode,
+    required bool? showNotification,
+    required Duration? seek,
+    required double? playSpeed,
+    required double? pitch,
+    required LoopMode? loopMode,
+    required HeadPhoneStrategy? headPhoneStrategy,
+    required AudioFocusStrategy? audioFocusStrategy,
+    required NotificationSettings? notificationSettings,
+    required bool? needRecord,
   }) async {
     final _autoStart = autoStart ?? _DEFAULT_AUTO_START;
     final _loopMode = loopMode ?? _DEFAULT_LOOP_MODE;
@@ -1050,11 +1063,13 @@ class AssetsAudioPlayer {
           'headPhoneStrategy': describeHeadPhoneStrategy(_headPhoneStrategy),
           'audioFocusStrategy': describeAudioFocusStrategy(_audioFocusStrategy),
           'displayNotification': _showNotification,
-          'volume': forcedVolume ?? volume.value ?? defaultVolume,
+          'volume': forcedVolume ?? volume.valueOrNull ?? defaultVolume,
           'playSpeed': playSpeed ??
               audio.playSpeed ??
-              this.playSpeed.value ??
+              this.playSpeed.valueOrNull ??
               defaultPlaySpeed,
+          'pitch':
+              pitch ?? audio.pitch ?? this.pitch.valueOrNull ?? defaultPitch,
         };
         if (seek != null) {
           params['seek'] = seek.inMilliseconds.round();
@@ -1067,6 +1082,15 @@ class AssetsAudioPlayer {
             audio.audioType == AudioType.liveStream) {
           params['networkHeaders'] =
               audio.networkHeaders ?? networkSettings.defaultHeaders;
+        }
+
+        if (audio.drmConfiguration != null) {
+          var drmMap = {};
+          drmMap['drmType'] = audio.drmConfiguration!.drmType.toString();
+          if (audio.drmConfiguration!.drmType == DrmType.clearKey) {
+            drmMap['clearKey'] = audio.drmConfiguration!.clearKey;
+          }
+          params['drmConfiguration'] = drmMap;
         }
 
         //region notifs
@@ -1136,12 +1160,13 @@ class AssetsAudioPlayer {
     bool showNotification = _DEFAULT_SHOW_NOTIFICATION,
     Duration? seek,
     double? playSpeed,
+    double? pitch,
     LoopMode? loopMode,
     NotificationSettings? notificationSettings,
     PlayInBackground? playInBackground,
     HeadPhoneStrategy headPhoneStrategy = _DEFAULT_HEADPHONE_STRATEGY,
     AudioFocusStrategy? audioFocusStrategy,
-    bool? needRecord,
+    bool needRecord = true,
   }) async {
     _lastSeek = null;
     _replaceRealtimeSubscription();
@@ -1151,6 +1176,7 @@ class AssetsAudioPlayer {
       respectSilentMode: respectSilentMode,
       showNotification: showNotification,
       playSpeed: playSpeed,
+      pitch: pitch,
       loopMode: loopMode,
       audioFocusStrategy: audioFocusStrategy ?? defaultFocusStrategy,
       notificationSettings: notificationSettings,
@@ -1191,6 +1217,7 @@ class AssetsAudioPlayer {
     bool showNotification = _DEFAULT_SHOW_NOTIFICATION,
     Duration? seek,
     double? playSpeed,
+    double? pitch,
     NotificationSettings? notificationSettings,
     LoopMode loopMode = _DEFAULT_LOOP_MODE,
     PlayInBackground playInBackground = _DEFAULT_PLAY_IN_BACKGROUND,
@@ -1219,22 +1246,21 @@ class AssetsAudioPlayer {
       }
 
       if (playlist != null) {
-        await _openPlaylist(
-          playlist,
-          autoStart: autoStart,
-          volume: volume,
-          respectSilentMode: respectSilentMode,
-          showNotification: showNotification,
-          seek: seek,
-          loopMode: loopMode,
-          playSpeed: playSpeed,
-          headPhoneStrategy: headPhoneStrategy,
-          audioFocusStrategy: focusStrategy,
-          notificationSettings:
-              notificationSettings ?? defaultNotificationSettings,
-          playInBackground: playInBackground,
-          needRecord: needRecord,
-        );
+        await _openPlaylist(playlist,
+            autoStart: autoStart,
+            volume: volume,
+            respectSilentMode: respectSilentMode,
+            showNotification: showNotification,
+            seek: seek,
+            loopMode: loopMode,
+            playSpeed: playSpeed,
+            pitch: pitch,
+            headPhoneStrategy: headPhoneStrategy,
+            audioFocusStrategy: focusStrategy,
+            notificationSettings:
+                notificationSettings ?? defaultNotificationSettings,
+            playInBackground: playInBackground,
+            needRecord: needRecord);
       }
       _acceptUserOpen = true;
     } catch (t) {
@@ -1250,8 +1276,7 @@ class AssetsAudioPlayer {
   ///     _assetsAudioPlayer.playOfPause();
   ///
   Future<void> playOrPause() async {
-    final playing = _isPlaying.value ?? true;
-    print("__interaption__playing $playing");
+    final playing = _isPlaying.valueOrNull ?? true;
     if (playing) {
       await pause();
     } else {
@@ -1355,11 +1380,11 @@ class AssetsAudioPlayer {
   ///
   Future<void> seekBy(Duration by) async {
     // only if playing a song
-    final playing = current.value;
+    final playing = current.valueOrNull;
     if (playing != null) {
       final totalDuration = playing.audio.duration;
 
-      final currentPosition = this.currentPosition.value ?? Duration();
+      final currentPosition = this.currentPosition.valueOrNull ?? Duration();
 
       if (by.inMilliseconds >= 0) {
         final nextPosition = currentPosition + by;
@@ -1373,7 +1398,7 @@ class AssetsAudioPlayer {
         await seek(currentPositionCapped);
       } else {
         // only if playing a song
-        final currentPosition = this.currentPosition.value ?? Duration();
+        final currentPosition = this.currentPosition.valueOrNull ?? Duration();
         final nextPosition = currentPosition + by;
 
         // don't seek less that 0
@@ -1429,6 +1454,22 @@ class AssetsAudioPlayer {
     });
   }
 
+  /// Change the current pitch of the MediaPlayer
+  ///
+  ///     _assetsAudioPlayer.setPitch(0.4);
+  ///
+  /// MIN : 0.0
+  /// MAX : 16.0
+  ///
+  /// if null, set to defaultPitch (1.0)
+  ///
+  Future<void> setPitch(double pitch) async {
+    await _sendChannel.invokeMethod('pitch', {
+      'id': id,
+      'pitch': pitch.clamp(minPitch, maxPitch),
+    });
+  }
+
   Future<Audio> _handlePlatformAsset(Audio input) async {
     if (defaultTargetPlatform == TargetPlatform.macOS &&
         input.audioType == AudioType.asset &&
@@ -1481,6 +1522,7 @@ class _CurrentPlaylist {
   final bool? showNotification;
   LoopMode? loopMode;
   final double? playSpeed;
+  final double? pitch;
   final NotificationSettings? notificationSettings;
   final AudioFocusStrategy? audioFocusStrategy;
   final PlayInBackground? playInBackground;
@@ -1582,6 +1624,7 @@ class _CurrentPlaylist {
     this.respectSilentMode,
     this.showNotification,
     this.playSpeed,
+    this.pitch,
     this.notificationSettings,
     this.playInBackground,
     this.loopMode,